:file_folder: Move boot -> app/boot
:art: Replaced boot/ with app/boot/ in files
