+++ app/boot/kernal/vars
@@ -1,0 +1,3 @@
+#!/usr/bin/bash
+echo Setting enviroment varibles...
+export WEB_HOME=/c/Users/Kishan Sambhi/Documents/Projects/Jakhu/app/boot/kernal
