+++ app/boot/libs/checks/checkFile.js
