+++ app/boot/src/vars.coffee
