+++ bin/jakhu
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 var express = require('express');
-var kernal = require('../boot/boot.js');
+var kernal = require('../app/boot/boot.js');
 //var git = require('nodegit');
 var exec = require('child_process').exec;
 var com = require('commander');
