+++ routes/dashboard.js
@@ -3,7 +3,7 @@ const express = require('express');
 let router = express.Router();
 const exec = require('child_process').exec;
 const delayed = require('delayed');
-const config = require('../boot/libs/configure.js');
+const config = require('../app/boot/libs/configure.js');
 const fs = require('fs');
 const YAML = require('yamljs');
 const runner = require('../libs/runner/findapp.js');
