+++ routes/index.js
@@ -4,7 +4,7 @@ var fs = require('fs');
 var dbName = "Jakhu";
 var port = "27017";
 var host = "localhost";
-var kernal = require('../boot/boot.js');
+var kernal = require('../app/boot/boot.js');
 var exec = require('child_process').exec;
 var xml2js = require('xml2js');
 var oobe = require('../libs/setup/setup.js');
