+++ test/boot/stop.js
@@ -1,5 +1,5 @@
 // test whether Jakhu can stop
-var stop = require('../../boot/index.js');
+var stop = require('../../app/boot/index.js');
 describe('Boot tests', function () {
   it('can we stop the server?', function (done) {
     // body...
